Add mutation for creating a terminal
Change upsert terminal to be just update terminal as Tiamat did not support upserting parent stop places
